+++ bevy_rtc/src/server/mod.rs
@@ -7,6 +7,6 @@ mod systems;
 
 pub use events::RtcServerEvent;
 pub use plugin::RtcServerPlugin;
-pub use router::AddProtocolExt;
+pub use router::AddServerProtocolExt;
 pub use state::{RtcServerState, RtcServerStatus};
 pub use system_params::RtcServer;