+++ demos/ping-server/src/main.rs
@@ -1,5 +1,5 @@
 use bevy::{log::LogPlugin, prelude::*};
-use bevy_rtc::server::{AddProtocolExt, RtcServer, RtcServerPlugin};
+use bevy_rtc::prelude::*;
 use protocol::PingPayload;
 
 fn main() {
@@ -7,7 +7,7 @@ fn main() {
         .add_plugins(MinimalPlugins)
         .add_plugins(LogPlugin::default())
         .add_plugins(RtcServerPlugin { port: 3536 })
-        .add_bounded_protocol::<PingPayload>(1)
+        .add_server_rw_protocol::<PingPayload>(1)
         .add_systems(Update, |mut server: RtcServer<PingPayload>| {
             for (peer_id, packet) in server.read() {
                 if let PingPayload::Ping = packet {