+++ demos/protocol/src/lib.rs
@@ -1,9 +1,9 @@
-use bevy_rtc::protocol::Payload;
+use bevy_rtc::protocol::Protocol;
 use serde::{Deserialize, Serialize};
 
 // Used by painting demo
 
-#[derive(Payload, Serialize, Deserialize, Debug, Clone)]
+#[derive(Protocol, Serialize, Deserialize, Debug, Clone)]
 pub struct DrawLinePayload {
     pub x1: f32,
     pub y1: f32,
@@ -11,7 +11,7 @@ pub struct DrawLinePayload {
     pub y2: f32,
 }
 
-#[derive(Payload, Serialize, Deserialize, Debug, Clone)]
+#[derive(Protocol, Serialize, Deserialize, Debug, Clone)]
 pub struct ChatPayload {
     pub from: String,
     pub message: String,
@@ -19,7 +19,7 @@ pub struct ChatPayload {
 
 // Used by ping demo
 
-#[derive(Payload, Serialize, Deserialize, Debug, Clone)]
+#[derive(Protocol, Serialize, Deserialize, Debug, Clone)]
 pub enum PingPayload {
     Ping,
     Pong,